+++ src/adapters/DOM.spec.ts
@@ -32,172 +32,131 @@ const isInvalid = (json: any) => {
 };
 
 describe('DOM adapter', () => {
-  it('sanity - should invalidate an img without alt', (done) => {
-    const options = {
-    };
-    const picture = makePicture(createElement, makePictureTree(handle, options));
-    validator({ data: makeHTML(picture) })
-      .then((data: any) => {
-        if (isInvalid(data)) {
-          done();
-        } else {
-          done('HTML was valid :(');
-        }
-      })
-      .catch((err: Error) => {
-        done(err);
-      });
+  it('sanity - should invalidate an img without alt', done => {
+    const options = {};
+    const picture = makePicture(
+      createElement,
+      makePictureTree(handle, options)
+    );
+    validator({ data: makeHTML(picture) }).then((data: any) => {
+      done(!isInvalid(data));
+    });
   });
 
-  it('should generate a valid picture element with width and pixel density', (done) => {
+  it('should generate a valid picture element with width and pixel density', done => {
     const options = {
       alt: 'downtown',
-      width: '768px',
+      width: '768px'
     };
-    const picture = makePicture(createElement, makePictureTree(handle, options));
-    validator({ data: makeHTML(picture) })
-      .then((data: any) => {
-        if (isInvalid(data)) {
-          done('HTML invalid');
-        } else {
-          done();
-        }
-      })
-      .catch((err: Error) => {
-        done(err);
-      });
+    const picture = makePicture(
+      createElement,
+      makePictureTree(handle, options)
+    );
+    validator({ data: makeHTML(picture) }).then((data: any) => {
+      done(isInvalid(data));
+    });
   });
 
-  it('should generate a valid picture element with 1 size and width descriptors', (done) => {
+  it('should generate a valid picture element with 1 size and width descriptors', done => {
     const options = {
       alt: 'downtown',
       sizes: {
         '(min-width: 640px)': '80vw',
-        fallback: '100vw',
+        fallback: '100vw'
       },
-      resolutions: [540, 670, 1080],
+      resolutions: [540, 670, 1080]
     };
-    const picture = makePicture(createElement, makePictureTree(handle, options));
-    validator({ data: makeHTML(picture) })
-      .then((data: any) => {
-        if (isInvalid(data)) {
-          done('HTML invalid');
-        } else {
-          done();
-        }
-      })
-      .catch((err: Error) => {
-        done(err);
-      });
+    const picture = makePicture(
+      createElement,
+      makePictureTree(handle, options)
+    );
+    validator({ data: makeHTML(picture) }).then((data: any) => {
+      done(isInvalid(data));
+    });
   });
 
-  it('should generate a valid picture element with 2 formats and 2 sizes', (done) => {
+  it('should generate a valid picture element with 2 formats and 2 sizes', done => {
     const options = {
       alt: 'downtown',
       sizes: {
         '(min-width: 640px)': '80vw',
         '(min-width: 320px)': '700px',
-        fallback: '100vw',
+        fallback: '100vw'
       },
-      formats: ['webp', 'jpg'],
+      formats: ['webp', 'jpg']
     };
-    const picture = makePicture(createElement, makePictureTree(handle, options));
-    validator({ data: makeHTML(picture) })
-      .then((data: any) => {
-        if (isInvalid(data)) {
-          done('HTML invalid');
-        } else {
-          done();
-        }
-      })
-      .catch((err: Error) => {
-        done(err);
-      });
+    const picture = makePicture(
+      createElement,
+      makePictureTree(handle, options)
+    );
+    validator({ data: makeHTML(picture) }).then((data: any) => {
+      done(isInvalid(data));
+    });
   });
 
-  it('should generate a valid picture element with 1 format and fallback', (done) => {
+  it('should generate a valid picture element with 1 format and fallback', done => {
     const options = {
       alt: 'downtown',
       sizes: {
-        fallback: '100vw',
+        fallback: '100vw'
       },
-      formats: ['webp', 'jpg'],
+      formats: ['webp', 'jpg']
     };
-    const picture = makePicture(createElement, makePictureTree(handle, options));
-    validator({ data: makeHTML(picture) })
-      .then((data: any) => {
-        if (isInvalid(data)) {
-          done('HTML invalid');
-        } else {
-          done();
-        }
-      })
-      .catch((err: Error) => {
-        done(err);
-      });
+    const picture = makePicture(
+      createElement,
+      makePictureTree(handle, options)
+    );
+    validator({ data: makeHTML(picture) }).then((data: any) => {
+      done(isInvalid(data));
+    });
   });
 
-  it('should generate a valid picture element with 1 format', (done) => {
+  it('should generate a valid picture element with 1 format', done => {
     const options = {
       alt: 'downtown',
-      formats: ['webp'],
+      formats: ['webp']
     };
-    const picture = makePicture(createElement, makePictureTree(handle, options));
-    validator({ data: makeHTML(picture) })
-      .then((data: any) => {
-        if (isInvalid(data)) {
-          done('HTML invalid');
-        } else {
-          done();
-        }
-      })
-      .catch((err: Error) => {
-        done(err);
-      });
+    const picture = makePicture(
+      createElement,
+      makePictureTree(handle, options)
+    );
+    validator({ data: makeHTML(picture) }).then((data: any) => {
+      done(isInvalid(data));
+    });
   });
 
-  it('should generate a valid picture element with img fallback size and 2 width descriptors', (done) => {
+  it('should generate a valid picture element with img fallback size and 2 width descriptors', done => {
     const options = {
       alt: 'downtown',
       sizes: {
-        fallback: '100vw',
+        fallback: '100vw'
       },
-      resolutions: ['320w', '640w'],
+      resolutions: ['320w', '640w']
     };
-    const picture = makePicture(createElement, makePictureTree(handle, options));
-    validator({ data: makeHTML(picture) })
-      .then((data: any) => {
-        if (isInvalid(data)) {
-          done('HTML invalid');
-        } else {
-          done();
-        }
-      })
-      .catch((err: Error) => {
-        done(err);
-      });
+    const picture = makePicture(
+      createElement,
+      makePictureTree(handle, options)
+    );
+    validator({ data: makeHTML(picture) }).then((data: any) => {
+      done(isInvalid(data));
+    });
   });
 
-  it('should generate a valid picture element with 1 fallback size and 1 format', (done) => {
+  it('should generate a valid picture element with 1 fallback size and 1 format', done => {
     const options = {
       alt: 'downtown',
       formats: ['webp'],
       resolutions: [640],
       sizes: {
-        fallback: '700px',
-      },
+        fallback: '700px'
+      }
     };
-    const picture = makePicture(createElement, makePictureTree(handle, options));
-    validator({ data: makeHTML(picture) })
-      .then((data: any) => {
-        if (isInvalid(data)) {
-          done('HTML invalid');
-        } else {
-          done();
-        }
-      })
-      .catch((err: Error) => {
-        done(err);
-      });
+    const picture = makePicture(
+      createElement,
+      makePictureTree(handle, options)
+    );
+    validator({ data: makeHTML(picture) }).then((data: any) => {
+      done(isInvalid(data));
+    });
   });
 });