+++ src/tree.ts
@@ -140,13 +140,22 @@ const createFileLink = (handle: FileHandle, fileLinkOptions: FileLinkOptions) =>
   if (!fileLinkOptions.useValidator || (fileLinkOptions.indexInSet && fileLinkOptions.indexInSet > 0)) {
     fileLink.setUseValidator(false);
   }
-  Object.keys(fileLinkOptions.transform).forEach((key: keyof TransformOptions) => {
+
+  Object.keys(fileLinkOptions.transform).sort(outputFirstSort).forEach((key: keyof TransformOptions) => {
     fileLink = fileLink.addTask(key, fileLinkOptions.transform[key]);
   });
   if (fileLinkOptions.cname) {
     fileLink.setCname(fileLinkOptions.cname);
   }
   return fileLink.toString();
+};
+
+/**
+ * Sort array of keys in a way that 'output' is always the first
+ * @param previousKey - First key to be compared in a sort function
+ */
+const outputFirstSort = (previousKey: string, nextKey: string) => {
+  return previousKey === 'output' ? -1 : nextKey === 'output' ? 1 : 0;
 };
 
 const getWidth = (width?: number | string) => (resolution: number | string) => {